fix(common/mocks): move `SimplifiedTrack` mocks to seperate file to avoid circular dependencies
